Moved all ModDatas to the ModData directory. Enhanced the Mod loading logics. Tried to add support to scripting in JavaScript.
