+++ src/ModLoader.cs
@@ -1,233 +1,178 @@
 #nullable enable
 using System;
 using System.Collections.Generic;
-using System.Drawing;
 using System.IO;
-using BepInEx;
 using BepInEx.Logging;
 using Newtonsoft.Json;
 
-namespace SFHR_ZModLoader 
+namespace SFHR_ZModLoader;
+
+public struct Mod
 {
-    [Serializable]
-    public struct ModMetadata
+    public ModMetadata metadata;
+    public Dictionary<string, ModNamespace> namespaces;
+    public Mod(ModMetadata metadata)
     {
-        public string id;
-        public string displayName;
-        public ulong versionCode;
-        public string version;
+        this.metadata = metadata;
+        this.namespaces = new();
     }
 
-    public struct ModNamespace 
+    public static Mod LoadFromDirectory(string dir, Mod? mod = null)
     {
-        public string name;
-        public Dictionary<string, ModCamoData> camoDatas;
-
-        public Dictionary<string, ModWeaponData> weaponDatas;
-
-        public static ModNamespace LoadFromDirectory(string dir, ModNamespace? ns = null)
+        if (!Directory.Exists(dir))
         {
-            if(!Directory.Exists(dir))
-            {
-                throw new ModLoadingException($"Namespace directory '{dir}' not exists.");
-            }
-            var nsname = Path.GetFileName(dir);
-            var camoDatas = new Dictionary<string, ModCamoData>();
-            var weaponDatas = new Dictionary<string, ModWeaponData>();
-            var nsConf = new ModNamespaceConf {
-                camos = "camos",
-                weapons = "weapons",
-            };
-            if(File.Exists(Path.Combine(dir, "namespace.json")))
-            {
-                var newConf = JsonConvert.DeserializeObject<ModNamespaceConf>(File.ReadAllText(Path.Combine(dir, "namespace.json")));
-                nsConf = new ModNamespaceConf {
-                    camos = newConf.camos ?? nsConf.camos,
-                    weapons = newConf.weapons ?? nsConf.weapons,
-                };
-            }
+            throw new ModLoadingException($"Mod directory '{dir}' not found.");
+        }
+        ModMetadata metadata;
+        try
+        {
+            metadata = JsonConvert.DeserializeObject<ModMetadata>(File.ReadAllText(Path.Combine(dir, "mod.json")));
+        }
+        catch
+        {
+            throw new ModLoadingException($"Errors in the metadata file 'mod.json'.");
+        }
+        var namespaces = mod?.namespaces ?? new Dictionary<string, ModNamespace>();
 
-            var camosConf = new ModCamosConf {};
-            if(File.Exists(Path.Combine(dir, nsConf.camos, "camos.json")))
+        foreach (var nsdir in Directory.EnumerateDirectories(dir))
+        {
+            if (namespaces.TryGetValue(Path.GetFileName(nsdir), out var ns))
             {
-                var newConf = JsonConvert.DeserializeObject<ModCamosConf>(File.ReadAllText(Path.Combine(dir, nsConf.camos, "camos.json")));
-                camosConf = newConf;
-            }
-            if(Directory.Exists(Path.Combine(dir, nsConf.camos)))
-            {
-                foreach (var item in Directory.EnumerateDirectories(Path.Combine(dir, nsConf.camos)))
-                {
-                    // TODO: includes 和 excludes处理
-                    var camoName = Path.GetFileName(item);
-                    if (ns?.camoDatas.TryGetValue(item, out var camoData) ?? false)
-                    {
-                        camoDatas[camoName] = ModCamoData.LoadFromDirectory(item, camoData);
-                    }
-                    else
-                    {
-                        camoDatas.Add(camoName, ModCamoData.LoadFromDirectory(item));
-                    }
-                }
+                namespaces[Path.GetFileName(nsdir)] = ModNamespace.LoadFromDirectory(nsdir, ns);
             }
             else
             {
-                SFHRZModLoaderPlugin.Logger?.LogInfo($"Skips camos at '{Path.Combine(dir, nsConf.camos)}'.");
+                namespaces.Add(Path.GetFileName(nsdir), ModNamespace.LoadFromDirectory(nsdir));
             }
-
-            var weaponsConf = new ModWeaponsConf {};
-            if(File.Exists(Path.Combine(dir, nsConf.weapons, "weapons.json")))
-            {
-                var newConf = JsonConvert.DeserializeObject<ModWeaponsConf>(File.ReadAllText(Path.Combine(dir, nsConf.weapons, "camos.json")));
-                weaponsConf = newConf;
-            }
-            if(Directory.Exists(Path.Combine(dir, nsConf.weapons)))
-            {
-                foreach (var item in Directory.EnumerateDirectories(Path.Combine(dir, nsConf.weapons)))
-                {
-                    // TODO: includes 和 excludes处理
-                    var weaponName = Path.GetFileName(item);
-                    if (ns?.weaponDatas.TryGetValue(item, out var weaponData) ?? false)
-                    {
-                        weaponDatas[weaponName] = ModWeaponData.LoadFromDirectory(item, weaponData);
-                    }
-                    else
-                    {
-                        weaponDatas.Add(weaponName, ModWeaponData.LoadFromDirectory(item));
-                    }
-                }
-            }
-            else
-            {
-                SFHRZModLoaderPlugin.Logger?.LogInfo($"Skips weapons at '{Path.Combine(dir, nsConf.weapons)}'.");
-            }
-            
-
-            return new ModNamespace {
-                name = nsname,
-                camoDatas = camoDatas,
-                weaponDatas = weaponDatas,
-            };
         }
-
-        public readonly void PatchToGameContext(GameContext gctx)
+        return new Mod
         {
-            foreach (var item in camoDatas)
-            {
-                item.Value.PatchToGameContext(gctx, name == "sfh" ? null : name);
-            }
-            foreach (var item in weaponDatas)
-            {
-                item.Value.PatchToGameContext(gctx, name == "sfh" ? null : name);
-            }
-        }
+            metadata = metadata,
+            namespaces = namespaces,
+        };
     }
 
-    public struct Mod {
-        public ModMetadata metadata;
-        public Dictionary<string, ModNamespace> namespaces;
-        public Mod(ModMetadata metadata)
-        {
-            this.metadata = metadata;
-            this.namespaces = new();
-        }
-
-        public static Mod LoadFromDirectory(string dir, Mod? mod = null)
-        {
-            if(!Directory.Exists(dir))
-            {
-                throw new ModLoadingException($"Mod directory '{dir}' not found.");
-            }
-            ModMetadata metadata;
-            try
-            {
-                metadata = JsonConvert.DeserializeObject<ModMetadata>(File.ReadAllText(Path.Combine(dir, "mod.json")));
-            }
-            catch
-            {
-                throw new ModLoadingException($"Errors in the metadata file 'mod.json'.");
-            }
-            var namespaces = mod?.namespaces ?? new Dictionary<string, ModNamespace>();
-
-            foreach (var nsdir in Directory.EnumerateDirectories(dir))
-            {
-                if(namespaces.TryGetValue(Path.GetFileName(nsdir), out var ns))
-                {
-                    namespaces[Path.GetFileName(nsdir)] = ModNamespace.LoadFromDirectory(nsdir, ns);
-                }
-                else
-                {
-                    namespaces.Add(Path.GetFileName(nsdir), ModNamespace.LoadFromDirectory(nsdir));
-                }
-            }
-            return new Mod {
-                metadata = metadata,
-                namespaces = namespaces,
-            };
-        }
-
-        public readonly void PatchToGameContext(GameContext gctx)
-        {
-            foreach (var item in namespaces)
-            {
-                item.Value.PatchToGameContext(gctx);
-            }
-        }
-    }
-
-    public class ModLoadingException: Exception
+    public readonly void PatchToGameContext(GameContext gctx)
     {
-        public ModLoadingException(string messages): base(messages)
-        {}
+        foreach (var item in namespaces)
+        {
+            item.Value.PatchToGameContext(gctx);
+        }
     }
 
-    public class ModLoader
+    public readonly void UnpatchToGameContext(GameContext gctx)
     {
-        private readonly string dir;
-        private Dictionary<string, Mod> mods;
-        private ManualLogSource? logger { get => SFHRZModLoaderPlugin.Logger; }
-
-        public ModLoader(string dir)
+        foreach (var item in namespaces)
         {
-            this.dir = dir;
-            this.mods = new();
-            logger?.LogInfo("ModLoader created.");
+            item.Value.UnpatchToGameContext(gctx);
         }
+    }
+}
 
-        public void RegisterEvents(EventManager eventManager)
+public class ModLoadingException : Exception
+{
+    public ModLoadingException(string messages) : base(messages)
+    { }
+}
+
+public class ModLoader
+{
+    private readonly string dir;
+    private Dictionary<string, Mod> mods;
+    private List<string>? modLoadOrder;
+    private ManualLogSource? logger { get => SFHRZModLoaderPlugin.Logger; }
+
+    public ModLoader(string dir)
+    {
+        this.dir = dir;
+        mods = new();
+        LoadModLoadOrder();
+        logger?.LogInfo("ModLoader created.");
+    }
+
+    public void LoadModLoadOrder()
+    {
+        if (File.Exists(Path.Combine(dir, "mod_load_order.txt")))
         {
-            eventManager.RegisterEventHandler("MODS_LOAD", ev => {
-                LoadMods();
-                eventManager.EmitEvent(new Event {
-                    type = "MODS_LOADED",
-                });
-            });
-            eventManager.RegisterEventHandler("GAMECONTEXT_PATCH", ev => {
-                if(ev.data == null || ev.data.GetType() != typeof(GameContext))
+            modLoadOrder = new();
+            foreach (var line in File.ReadLines(Path.Combine(dir, "mod_load_order.txt")))
+            {
+                if (!line.TrimStart().StartsWith("#"))
                 {
-                    logger?.LogError("GAMECONTEXT_PATCH data incorrect!");
-                    return;
+                    modLoadOrder.Add(line.Trim());
                 }
-                LoadMods();
-                var gctx = (GameContext)ev.data;
-                logger?.LogInfo("Game patching...");
-                PatchToGameContext(gctx);
-                logger?.LogInfo("Game patch completed.");
-            });
-            eventManager.RegisterEventHandler("GAMECONTEXT_LOADED", ev => {
-                var gctx = (GameContext)ev.data;
-                eventManager.EmitEvent(new Event {
-                    type = "GAMECONTEXT_PATCH", 
-                    data = gctx,
-                });
-            });
-        }
-
-        public void LoadMods()
-        {
-            logger?.LogInfo($"Loading Mods from directory: {dir}...");
-            if(!Directory.Exists(dir)) {
-                Directory.CreateDirectory(dir);
             }
+            logger?.LogInfo("'mod_load_order.txt' loaded.");
+        }
+        else
+        {
+            modLoadOrder = null;
+            logger?.LogInfo("Skipped 'mod_load_order.txt' because it is missing, defaults to load all the mods.");
+        }
+    }
+
+    public void RegisterEvents(EventManager eventManager)
+    {
+        eventManager.RegisterEventHandler("MODS_LOAD", ev =>
+        {
+            LoadMods();
+            eventManager.EmitEvent(new Event
+            {
+                type = "MODS_LOADED",
+            });
+        });
+        eventManager.RegisterEventHandler("GAMECONTEXT_PATCH", ev =>
+        {
+            if (ev.data == null || ev.data.GetType() != typeof(GameContext))
+            {
+                logger?.LogError("GAMECONTEXT_PATCH data incorrect!");
+                return;
+            }
+            LoadMods();
+            var gctx = (GameContext)ev.data;
+            logger?.LogInfo("Game patching...");
+            PatchToGameContext(gctx);
+            logger?.LogInfo("Game patch completed.");
+        });
+        eventManager.RegisterEventHandler("GAMECONTEXT_LOADED", ev =>
+        {
+            var gctx = (GameContext)ev.data;
+            eventManager.EmitEvent(new Event
+            {
+                type = "GAMECONTEXT_PATCH",
+                data = gctx,
+            });
+        });
+        eventManager.RegisterEventHandler("MODS_RELOAD", ev =>
+        {
+            if (SFHRZModLoaderPlugin.GameContext != null)
+            {
+                UnpatchToGameContext(SFHRZModLoaderPlugin.GameContext);
+            }
+            UnloadMods();
+            LoadModLoadOrder();
+            LoadMods();
+            if (SFHRZModLoaderPlugin.GameContext != null)
+            {
+                eventManager.EmitEvent(new Event
+                {
+                    type = "GAMECONTEXT_PATCH",
+                    data = SFHRZModLoaderPlugin.GameContext,
+                });
+            }
+        });
+        logger?.LogInfo("All ModLoader events registered.");
+    }
+
+    public void LoadMods()
+    {
+        logger?.LogInfo($"Loading Mods from directory: {dir}...");
+        if (!Directory.Exists(dir))
+        {
+            Directory.CreateDirectory(dir);
+        }
+        if (modLoadOrder == null)
+        {
             foreach (var item in Directory.EnumerateDirectories(dir))
             {
                 if (File.Exists(Path.Combine(item, "mod.json")))
@@ -236,7 +181,7 @@ namespace SFHR_ZModLoader
                     try
                     {
                         logger?.LogInfo($"Loading Mod from directory: {item}...");
-                        if(mods.TryGetValue(metadata.id, out var mod))
+                        if (mods.TryGetValue(metadata.id, out var mod))
                         {
                             this.mods[mod.metadata.id] = Mod.LoadFromDirectory(item, mod);
                         }
@@ -246,32 +191,76 @@ namespace SFHR_ZModLoader
                         }
                         logger?.LogInfo($"Loading Mod '{metadata.id}' completed.");
                     }
-                    catch(Exception e)
+                    catch (Exception e)
                     {
                         logger?.LogError($"Load Mod in '{item}' failed: {e}.");
                     }
                 }
             }
         }
-
-        public Mod? GetMod(string name)
+        else
         {
-            if (mods.ContainsKey(name))
+            foreach (var modName in modLoadOrder)
             {
-                return mods[name];
-            }
-            else
-            {
-                return null;
+                if (File.Exists(Path.Combine(dir, modName, "mod.json")))
+                {
+                    var metadata = JsonConvert.DeserializeObject<ModMetadata>(File.ReadAllText(Path.Combine(dir, modName, "mod.json")));
+                    try
+                    {
+                        logger?.LogInfo($"Loading Mod from directory: {Path.Combine(dir, modName)}...");
+                        if (mods.TryGetValue(metadata.id, out var mod))
+                        {
+                            this.mods[mod.metadata.id] = Mod.LoadFromDirectory(Path.Combine(dir, modName), mod);
+                        }
+                        else
+                        {
+                            this.mods.Add(metadata.id, Mod.LoadFromDirectory(Path.Combine(dir, modName)));
+                        }
+                        logger?.LogInfo($"Loading Mod '{metadata.id}' completed.");
+                    }
+                    catch (Exception e)
+                    {
+                        logger?.LogError($"Load Mod in '{Path.Combine(dir, modName)}' failed: {e}.");
+                    }
+                }
+                else
+                {
+                    logger?.LogWarning($"Skipped load Mod '{Path.Combine(dir, modName)}': Not exist.");
+                }
             }
         }
+    }
 
-        public void PatchToGameContext(GameContext gctx)
+    public void UnloadMods()
+    {
+
+    }
+
+    public Mod? GetMod(string name)
+    {
+        if (mods.ContainsKey(name))
         {
-            foreach (var item in mods)
-            {
-                item.Value.PatchToGameContext(gctx);
-            }
+            return mods[name];
+        }
+        else
+        {
+            return null;
+        }
+    }
+
+    public void PatchToGameContext(GameContext gctx)
+    {
+        foreach (var item in mods)
+        {
+            item.Value.PatchToGameContext(gctx);
+        }
+    }
+
+    public void UnpatchToGameContext(GameContext gctx)
+    {
+        foreach (var item in mods)
+        {
+            item.Value.UnpatchToGameContext(gctx);
         }
     }
 }
